requirements.txt
# Core Dependencies
streamlit>=1.31.0
python-dotenv>=1.0.0

# PDF Processing
pypdf>=4.0.0
PyPDF2>=3.0.0

# LangChain Components
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.11
langchain-huggingface>=0.0.1

# Vector Store & Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.3.0

# HuggingFace
huggingface-hub>=0.20.0
transformers>=4.36.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0